Adapt to Bankscrap 2 API
