+++ lib/bankscrap/openbank/card.rb
@@ -5,7 +5,7 @@ module Bankscrap
     class Card < Account
       include Utils
 
-      CARD_ENDPOINT = '/OPB_BAMOBI_WS_ENS/ws/BAMOBI_WS_Def_Listener'
+      CARD_ENDPOINT = '/OPB_BAMOBI_WS_ENS/ws/BAMOBI_WS_Def_Listener'.freeze
 
       # Fetch transactions for the given account.
       # By default it fetches transactions for the last month,
@@ -23,7 +23,7 @@ module Bankscrap
           transactions += document.xpath('//lista/dato').map { |data| build_transaction(data) }
 
           repos = document.at_xpath('//methodResult/repos')
-          end_page = !(value_at_xpath(document, '//methodResult/finLista') == 'N')
+          end_page = value_at_xpath(document, '//methodResult/finLista') != 'N'
         end
 
         transactions
@@ -39,7 +39,7 @@ module Bankscrap
         <soapenv:Body>
           <v1:listaMovTarjetasFechas_LIP facade="BAMOBITAJ">
             <entrada>
-              <datosConexion>#{connection.user_data.children.to_s}</datosConexion>
+              <datosConexion>#{connection.user_data.children}</datosConexion>
               #{connection.xml_datos_cabecera}
               <contratoTarjeta>#{contract_id}</contratoTarjeta>
               <numeroTarj>#{id}</numeroTarj>
@@ -60,16 +60,14 @@ module Bankscrap
 
       # Build a transaction object from API data
       def build_transaction(data)
-        currency = value_at_xpath(data, 'importeMovto/DIVISA')
         Transaction.new(
           account: self,
           id: value_at_xpath(data, 'movimDia'),
-          amount: money(value_at_xpath(data, 'importeMovto/IMPORTE'), currency),
+          amount: money(data, 'importeMovto'),
           description: value_at_xpath(data, 'descMovimiento'),
-          effective_date: Date.strptime(value_at_xpath(data, 'fechaAnota'), "%Y-%m-%d"),
-          #operation_date: Date.strptime(value_at_xpath(data, 'fechaOpera'), "%Y-%m-%d"),
-          currency: currency,
-          balance: nil
+          effective_date: Date.strptime(value_at_xpath(data, 'fechaAnota'), '%Y-%m-%d'),
+          # operation_date: Date.strptime(value_at_xpath(data, 'fechaOpera'), "%Y-%m-%d"),
+          balance: Money.new(0, 'EUR') # TODO: Prepaid/debit cards don't have a Balance - maybe Credit ones do.
         )
       end
     end
